backend/pyproject.toml
[tool.poetry]
name = "kedb-platform"
version = "0.1.0"
description = "Agentic Known Error Database with cited recommendations, governance, and analytics."
authors = ["JagrukOps <ops@example.com>"]
license = "MIT"
readme = "../README.md"
packages = [
  {include = "app"}
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.30"
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
rq = "^1.16.2"
meilisearch = "^0.30.0"
httpx = "^0.27.0"
structlog = "^24.2.0"
orjson = "^3.10.0"
tenacity = "^8.3.0"
openai = "^1.30.1"
pydantic = "^2.7.1"
pgvector = "^0.2.5"
bcrypt = "^4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.6"
anyio = "^4.3.0"
ruff = "^0.4.5"

[tool.poetry.scripts]
kedb-api = "app.main:run"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
